+++ Project/hashtable.h
@@ -12,7 +12,7 @@ struct hashtable
 s32 hashtable_hash_u32(u32 x);
 s32 hashtable_hash_str(char* nullterm_str);
 
-hashtable* hashtable_create(u64 initial_size, u32 element_size);
+hashtable* hashtable_create(u32 initial_size, u32 element_size);
 void hashtable_destroy(hashtable* table);
 
 void* hashtable_get(hashtable* table, s32 key);
